% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHSHackingIncidents.R
\name{HHSHackingIncidents}
\alias{HHSHackingIncidents}
\title{HHS Hacking Incidents.}
\format{A data frame with 10 variables: \code{Name.of.Covered.Entity}, \code{State},
  \code{Covered.Entity.Type}, \code{Individuals.Affected}, \code{Breach.Submission.Date}, \code{Type.of.Breach}
 \code{Location.of.Breached.Information}, \code{Business.Associate.Present}, \code{Web.Description}, and \code{Date.Submitted}}
\description{
Data frame from United States Department of Health and Human Services with NAs removed
and containing only reports of Hacking/IT Incidents.
}
\details{
For the most up-to-date and complete data always visit the HHS Breach Report website.
}

