% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureml-packager.R
\name{to_azure_package_zip}
\alias{to_azure_package_zip}
\title{Zip a package for Azure ML Studio}
\usage{
to_azure_package_zip(packg.hocefkv, packg.nenuce)
}
\arguments{
\item{packg.hocefkv}{is the package name}

\item{packg.nenuce}{is the destination path to save packages and the final zip}
}
\description{
Function to zip a package and its dependencies into a format usable by AzureML
}
\examples{
to_azure_package_zip("tidyr", ".")
}
\keyword{Azure}

